Split tests into separate source files
